Move DeletedResourceTest inside SimpleProjectTest
